Move backend automation tests into a separate module
